apps/backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
psycopg[binary]==3.2.9
pydantic==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.20
PyMuPDF==1.26.3
cryptography==45.0.6
scikit-learn==1.5.2